% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_utf8.R
\name{collect}
\alias{collect}
\title{A variant of dplyr's collect() which converts dataframes to UTF-8 encoding
if OS is Windows}
\usage{
collect(x)
}
\arguments{
\item{x}{A tbl}
}
\description{
Works as a simple \code{\link[dplyr:compute]{collect()}}, on which it is
based.
However, on a Windows OS the \code{collect()} function exported
by this package will convert character and factor
variables of dataframes to \code{UTF-8} encoding.
}
\details{
The function \code{\link[=convertdf_enc]{convertdf_enc()}} is the workhorse
for the encoding conversion.
}
