% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_utf8.R
\docType{import}
\name{convertdf_enc}
\alias{convertdf_enc}
\title{Convert encoding of character and factor variables in a dataframe}
\keyword{documentation}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{inborutils}{\code{\link[inborutils]{convertdf_enc}}}
}}

