% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_watina.R
\name{dbDisconnect}
\alias{dbDisconnect}
\title{Disconnect a database connection}
\description{
This is a re-export of
\code{\link[inbodb:dbDisconnect-OdbcConnection-method]{inbodb::dbDisconnect()}}
(\href{https://inbo.github.io/inbodb/reference/dbDisconnect-OdbcConnection-method.html}{url}).
}
\keyword{documentation}
