% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{eval_chem}
\alias{eval_chem}
\title{Evaluate hydrochemical data per location}
\usage{
eval_chem(
  data,
  chem_var = c("P-PO4", "N-NO3", "N-NO2", "N-NH4", "HCO3", "SO4", "Cl", "Na", "K", "Ca",
    "Mg", "Fe", "Mn", "Si", "Al", "CondF", "CondL", "pHF", "pHL"),
  type = c("avail", "num", "both"),
  uniformity_test = FALSE
)
}
\arguments{
\item{data}{An object returned by \code{\link{get_chem}}.}

\item{chem_var}{A character vector to select chemical variables for which
statistics will be computed.
To specify chemical variables, use the
codes from the column \code{chem_variable} in \code{data}.}

\item{type}{A string defining the requested type of summary statistics.
See section 'Value'.
Either:
\itemize{
\item{\code{"avail"}:} availability statistics (the default);
\item{\code{"num"}:} numeric summary statistics;
\item{\code{"both"}:} both types will be returned.
}}

\item{uniformity_test}{Should the availability statistic
\code{pval_uniform_totalspan} be added (see section 'Value')?
Defaults to \code{FALSE} as this takes much more time to calculate than
everything else.}
}
\value{
A tibble with variables \code{loc_code} (see \code{\link{get_locs}}),
\code{chem_variable} (character; see Details)
and summary statistics (\emph{at the level
of \code{loc_code} x \code{chem_variable}}) depending on the state of
\code{type}
(\code{type = "both"} combines both cases):

With \code{type = "avail"} (the default):
\itemize{
\item{\code{nryears}}: number of calendar years for which the
chemical variable is available (on one date at least)
\item{\code{nrdates}}: number of dates at which the
chemical variable is available
\item{\code{firstdate}}: earliest date at which the
chemical variable is available
\item{\code{lastdate}}: latest date at which the
chemical variable is available
\item{\code{timespan_years}}: duration as years from the first calendar year
(year of \code{firstdate}) to the last calendar year
(year of \code{lastdate}) with data available
\item{\code{timespan_totalspan_ratio}}: the ratio of \code{timespan_years} to
the 'total span', which is the duration as years from the first to the last
calendar year \emph{for the whole of \strong{\code{data}}}.
\item{\code{nryears_totalspan_ratio}}: the ratio of \code{nryears} to
the 'total span'
\item{\code{pval_uniform_totalspan}}: Only returned with
\code{uniformity_test = TRUE}.
p-value of an exact, one-sample two-sided
Kolmogorov-Smirnov test for the discrete uniform distribution of the calendar
years \emph{with data of the chemical variable available} within the series
of consecutive calendar years defined by the 'total span' (see above).
The smaller the p-value,
the less uniform the years are spread within the total span.
A perfectly uniform spread results in a p-value of 1.
}

With \code{type = "num"}: summary statistics based on the chemical values,
i.e. excluding the 'combined' level:
\itemize{
\item{\code{val_min}}: minimum value
\item{\code{val_pct10}, \code{val_pct25}, \code{val_pct50},
\code{val_pct75}, \code{val_pct90}}: percentiles
\item{\code{val_max}}: maximum value
\item{\code{val_range}}: range
\item{\code{val_mean}}: mean
\item{\code{val_geometric_mean}}: geometric mean.
Only calculated when
all values are strictly positive.
\item{\code{unit}}
\item{\code{prop_below_loq}}: the proportion of measurements below the limit
of quantification (as derived from \code{below_loq == TRUE}),
i.e. relative to the total number of measurements
\emph{for which \code{below_loq} is not \code{NA}}.
This statistic neglects measurements with value \code{NA} for
\code{below_loq}!
If all measurements have value \code{NA} for
\code{below_loq}, this statistic is set to \code{NA}.
}
}
\description{
For a dataset as returned by \code{\link{get_chem}},
return summary statistics (data availability and/or
numeric properties) of
the specified hydrochemical variables, for each location.
}
\details{
For the availability statistics, an extra level
"\code{combined}" is added in the column \code{chem_variable} whenever the
arguments \code{data} and
\code{chem_var} imply more than one
chemical variable to be investigated.
This 'combined' level defines data availability for a water sample as the
availability of data for \strong{all} corresponding chemical variables.
}
\examples{
\dontrun{
watina <- connect_watina()
library(dplyr)
mylocs <- get_locs(watina, area_codes = "ZWA")
mydata <-
  mylocs \%>\%
  get_chem(watina, "1/1/2010")
mydata \%>\% arrange(loc_code, date, chem_variable)
mydata \%>\%
  pull(date) \%>\%
  lubridate::year(.) \%>\%
  (function(x) c(firstyear = min(x), lastyear = max(x)))
mydata \%>\%
  eval_chem(chem_var = c("P-PO4", "N-NO3", "N-NO2", "N-NH4")) \%>\%
  arrange(desc(loc_code))
mydata \%>\%
  eval_chem(
    chem_var = c("P-PO4", "N-NO3", "N-NO2", "N-NH4"),
    type = "both"
  ) \%>\%
  arrange(desc(loc_code)) \%>\%
  as.data.frame() \%>\%
  head(10)
mydata \%>\%
  eval_chem(
    chem_var = c("P-PO4", "N-NO3", "N-NO2", "N-NH4"),
    uniformity_test = TRUE
  ) \%>\%
  arrange(desc(loc_code)) \%>\%
  select(loc_code, chem_variable, pval_uniform_totalspan)
# Disconnect:
dbDisconnect(watina)
}

}
\seealso{
Other functions to evaluate locations: 
\code{\link{eval_xg3_avail}()},
\code{\link{eval_xg3_series}()}
}
\concept{functions to evaluate locations}
