% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{filter_xg3}
\alias{filter_xg3}
\title{Filter XG3 data per location x hydroyear}
\usage{
filter_xg3(data, xg3_type)
}
\arguments{
\item{data}{An object returned by \code{\link{get_xg3}}.}

\item{xg3_type}{Character vector of length 1, 2 or 3.
Defines the types of XG3 which are taken from \code{data}.
Specifies the 'X' in 'XG3': either \code{"L"}, \code{"H"} and/or \code{"V"}.}
}
\value{
A \code{tbl_lazy} object or a tibble, which is like \code{data} but with
non-requested XG3 variables discarded.
}
\description{
Helper function to select the specified XG3 columns per location x hydroyear.
}
\keyword{internal}
