% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{qualify_xg3}
\alias{qualify_xg3}
\title{Qualify XG3 data per location x hydroyear}
\usage{
qualify_xg3(data, xg3_type, combined)
}
\arguments{
\item{data}{An object returned by \code{\link{get_xg3}}.}

\item{xg3_type}{Character vector of length 1, 2 or 3.
Defines the types of XG3 which are taken to evaluate the hydroyears for
each location: either \code{"LG3"}, \code{"HG3"} and/or \code{"VG3"}.}

\item{combined}{Logical.
If \code{xg3_type} has length two or three:
for a hydroyear to be evaluated as 'XG3 is available',
should \emph{all} selected XG3 types be non-\code{NA}
(\code{combined = TRUE}),
or \emph{at least one} (\code{combined = FALSE})?}
}
\value{
A \code{tbl_lazy} object or a tibble, with a variable \code{meets_condition}
to denote whether the hydroyear has the requested XG3 values available,
at each location.
}
\description{
Helper function to select the specified XG3 columns and determine
if availability-condition is met per location x hydroyear.
}
\keyword{internal}
