% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chem.R
\encoding{UTF-8}
\name{read_vanwirdum_data}
\alias{read_vanwirdum_data}
\title{Return the 'vanwirdum_data' data source as a tibble}
\usage{
read_vanwirdum_data()
}
\value{
The \code{vanwirdum_data} dataframe as a
\code{\link[tibble:tbl_df-class]{tibble}}
}
\description{
Returns the included data source 'vanwirdum_data' as a
\code{\link[tibble:tbl_df-class]{tibble}}
}
\details{
Source: Van Wirdum, Geert (1991). Vegetation and hydrology of floating
rich-fens. Datawyse, Maastricht. 316 p. ISBN 90-5291-045-6. (Appendix D)
\href{https://publicwiki.deltares.nl/display/VWD/Home}{dataset available here}

The 'vanwirdum_data' dataset gives the curved contour which encloses the
plotting area of all possible, simple mixtures of the following reference
water samples:

- LI-ANG (a relatively calcium-rich groundwater sample),

- AT-WTV (a precipitation sample caught in a relatively unpolluted inland
area of The Netherlands)

- TH-N70 (a representative analysis from the North Sea
monitoring program, 70 km from the coast).

The curved contour around these points gives the boundary within which
most natural waters can occur.

\code{read_vanwirdum_data()} returns this dataset as a
\code{\link[tibble:tbl_df-class]{tibble}} with 42 rows and 2 variables.
A tibble is a dataframe that makes working in the tidyverse a little
\href{https://r4ds.had.co.nz/tibbles.html}{easier}.

Variables:
\describe{
  \item{ec25}{electric conductivity at 25°C in mS/m}
  \item{ir}{ionic ratio in \%}}
}
\examples{
read_vanwirdum_data()

}
\keyword{internal}
