% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectlocs.R
\name{selectlocs}
\alias{selectlocs}
\title{Select locations based on summary statistics (generic function)}
\usage{
selectlocs(
  data,
  data_type = c("data", "summary"),
  eval_fun = NULL,
  eval_args = NULL,
  conditions,
  verbose,
  list
)
}
\arguments{
\item{data}{Either:
\itemize{
\item{a dataset (data frame or lazy object) with the variable's values}:
with at least a column \code{loc_code}, a column with the variable name,
and a column \code{value};
\item{a data frame with summary statistics}:
with a first column \code{loc_code} and a second column with the evaluated
variable names or codes.
The column name of the second column can vary.
All other columns should have the name of the summary statistic, and hold
its value for \code{loc_code} x variable.
}}

\item{data_type}{A string.
Either \code{"data"} (the default) or \code{"summary"}, in correspondence
with the choice made for \code{data}.}

\item{eval_fun}{The evaluation function to be run, if \code{data} is a
dataset.}

\item{eval_args}{The arguments of the evaluation function, as a named list,
if \code{data} is a dataset.}

\item{conditions}{A data frame.
It must have the
following columns:
\describe{
\item{\code{variable}}{Can be any variable,
including \code{"combined"}.}
\item{\code{statistic}}{Name of the statistic to be evaluated.}
\item{\code{criterion}}{Numeric. Defines the value of the statistic on which
the
condition will be based.}
\item{\code{direction}}{One of: \code{"min","max","equal"}.
Together with \code{criterion}, this completes the condition which will
be evaluated with respect to the specific \code{chem_variable}:
for \code{direction = "min"}, the statistic must be the criterion
value or larger; for \code{direction = "max"}, the statistic must be
the criterion value or lower; for \code{direction = "equal"},
the statistic must be equal to the criterion value.
}
}

Each condition is one row of the data frame.
The data frame should have at least one, and may have many.
Each combination of \code{chem_variable} and \code{statistic} must be
unique.
Conditions on chemical variables, absent from \code{data} or not implied by
\code{chem_var}, will be dropped without warning.
Hence, it is up to the user to do sensible things.}

\item{verbose}{Logical.
If \code{TRUE}, give feedback on dropped locations because of
(specific) unused conditions and other 'mismatch' reasons.}

\item{list}{Logical.
If \code{FALSE} (the default), the function only returns the end-result
(a tibble with selected location codes).
If \code{TRUE}, the function returns a list with the end-result plus useful
intermediate results.}
}
\value{
If \code{list = FALSE}: a tibble with one column \code{loc_code} that
provides the locations selected by the conditions.

If \code{list = TRUE}: a list of tibbles that extends the previous end-result
with intermediate results.
}
\description{
Generic (helper) function to select locations that comply with
user-specified conditions,
from either a dataset (data frame or lazy object) with the variable's values or
from a data frame with summary statistics.
}
\details{
The result of the evaluation function (\code{eval_fun}) must produce a
data frame, formatted as declared by the second bullet under the
\code{data} argument.
}
\keyword{internal}
