% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{warn_xy_duplicates}
\alias{warn_xy_duplicates}
\title{Warn for duplicated XY coordinate pairs}
\usage{
warn_xy_duplicates(x, y)
}
\arguments{
\item{x}{Numerical vector of x coordinates}

\item{y}{Numerical vector of y coordinates}
}
\description{
Warn for duplicated XY coordinate pairs
}
\details{
Note that both vectors must be of same length.
}
\keyword{internal}
