% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watina.R
\docType{package}
\name{watina}
\alias{watina}
\alias{watina-package}
\title{\code{watina} package}
\description{
Querying & preprocessing data from the INBO Watina database (mainly groundwater data)
}
\details{
See the README on
\href{https://github.com/inbo/watina/blob/master/README.md}{GitHub}.
}
