% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_chem.R
\name{calc_ir}
\alias{calc_ir}
\title{Compute the ionic ratio (e.g. for a Van Wirdum diagram)}
\usage{
calc_ir(x)
}
\arguments{
\item{x}{Output of the get_chem function (hydrochemical data retrieved from
Watina). It can be a lazy object as well as a local object (get_chem(collect = TRUE))}
}
\value{
A tibble similar to x but with an extra column ir with the ionic ratio
}
\description{
Uses the output of get_chem and convert it to a new tibble with the ionic ratio
(and the right input format to make a Van Wirdum diagram).
}
\details{
Depending on the units of the Ca and Cl concentrations, the ionic ratio gets
calculated as:

1) with the Ca and Cl concentrations in meq/l

ionic ratio = [Ca2+] / ([Ca2+] + [Cl-])

2) with the Ca and Cl concentrations in mg/l

ionic ratio = ((Ca_mg*2)/40.078) / (((Ca_mg*2)/40.078) + (Cl_mg/35.453))

The result is the ionic ratio ir without units (0-1).
}
\examples{
\dontrun{
  watina <- connect_watina()

  # get the chemical data
  mydata <-
  get_locs(watina, area_codes = "ZWA") \%>\%
  get_chem(watina, "1/1/2019") \%>\%
  collect \%>\%
  as.data.frame

  # compute ionic ratio and add as new field
  mydata_vw <- calc_ir(mydata)
  # this dataset can be used to make a Van Wirdum diagram with ggplot_vanwirdum_background()

}

}
