% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_chem.R
\name{calculate_ir}
\alias{calculate_ir}
\title{Compute the ionic ratio}
\usage{
calculate_ir(data)
}
\arguments{
\item{data}{Output of the \code{get_chem} function (hydrochemical data
retrieved from Watina) or any dataset with at least the columns
\code{loc_code},
\code{chem_variable} (with at least concentrations for Ca and Cl),
\code{value}, \code{unit} and \code{date}.
It can be a lazy object as well as a local object
\code{(get_chem(collect = TRUE))}.}
}
\value{
A tibble similar to the input \code{data} but with an extra column \code{ir}
with the calculated ionic ratio.
}
\description{
Calculate the ionic ratio of a water sample based on the output of
\code{\link{get_chem}} and add it to the input data as a new column \code{ir}.
}
\details{
Depending on the units of the Ca and Cl concentrations, the ionic ratio gets
calculated as follows:

\enumerate{
 \item{if the Ca and Cl concentrations are in meq/l:

 ionic ratio = [Ca2+] / ([Ca2+] + [Cl-])}
 \item{if the Ca and Cl concentrations are in mg/l:

 ionic ratio = ((Ca_mg*2)/40.078) / (((Ca_mg*2)/40.078) + (Cl_mg/35.453))}
}

The result is the ionic ratio \code{ir} without units (0-1).
}
\examples{
\dontrun{
watina <- connect_watina()

# get the chemical data
mydata <-
  get_locs(watina, area_codes = "ZWA") \%>\%
  get_chem(watina, "1/1/2019") \%>\%
  collect()

# compute ionic ratio and add as new field
mydata_with_ir <- calculate_ir(mydata)
}

}
