% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_watina.R
\name{connect_watina}
\alias{connect_watina}
\title{Connect to the INBO Watina database}
\usage{
connect_watina()
}
\value{
A \code{DBIConnection} object.
}
\description{
Returns a connection to the INBO \strong{Watina} database.
The function can only be used from within the INBO network.
}
\details{
Don't forget to disconnect at the end of your R-script using
\code{\link{dbDisconnect}}!
}
\examples{
\dontrun{
watina <- connect_watina()
# Do your stuff.
# Disconnect:
dbDisconnect(watina)
}

}
