% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{eval_xg3_avail}
\alias{eval_xg3_avail}
\title{Evaluate the availability of XG3 values per location}
\usage{
eval_xg3_avail(data, xg3_type = c("L", "H", "V"), combined = TRUE)
}
\arguments{
\item{data}{An object returned by \code{\link{get_xg3}}.}

\item{xg3_type}{Character vector of length 1, 2 or 3.
Defines the types of XG3 which are taken to evaluate the hydroyears for
each location: either \code{"LG3"}, \code{"HG3"} and/or \code{"VG3"}.}

\item{combined}{Logical.
If \code{xg3_type} has length two or three:
for a hydroyear to be evaluated as 'XG3 is available',
should \emph{all} selected XG3 types be non-\code{NA}
(\code{combined = TRUE}),
or \emph{at least one} (\code{combined = FALSE})?}
}
\value{
A tibble with variables \code{loc_code} (see \code{\link{get_locs}}),
\code{nryears}, \code{firstyear} and \code{lastyear}.
}
\description{
For a dataset as returned by \code{\link{get_xg3}},
return for each location the first and
last (hydro)year and the number of (hydro)years with available XG3 values
of the specified type(s) (LG3, HG3 and/or VG3).
}
\examples{
\dontrun{
watina <- connect_watina()
library(dplyr)
mylocs <- get_locs(watina, area_codes = "KAL")
mydata <-
 mylocs \%>\%
 get_xg3(watina, 2014)
mydata
eval_xg3_avail(mydata,
               xg3_type = c("L", "V"))
eval_xg3_avail(mydata,
               xg3_type = c("L", "V"),
               combined = FALSE)
# Disconnect:
DBI::dbDisconnect(watina)
}


}
