% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{eval_xg3_avail}
\alias{eval_xg3_avail}
\title{Evaluate the availability of XG3 values per location}
\usage{
eval_xg3_avail(data, xg3_type = c("L", "H", "V"))
}
\arguments{
\item{data}{An object returned by \code{\link{get_xg3}}.}

\item{xg3_type}{Character vector of length 1, 2 or 3.
Defines the types of XG3 which are taken to evaluate the hydroyears for
each location.
Specifies the 'X' in 'XG3': either \code{"L"}, \code{"H"} and/or \code{"V"}.}
}
\value{
A tibble with variables \code{loc_code} (see \code{\link{get_locs}}),
\code{xg3_variable} (character; see Details),
\code{nryears}, \code{firstyear} and \code{lastyear}.
}
\description{
For a dataset as returned by \code{\link{get_xg3}},
return for each location the first and
last (hydro)year and the number of (hydro)years with available XG3 values
of the specified type(s) (LG3, HG3 and/or VG3).
}
\details{
The column \code{xg3_variable} in the resulting tibble
lists the requested XG3 types, for each location.
If more than one XG3 type is requested, or if \code{data} contains both
vertical CRSes, an extra value "\code{combined}" is listed in the
\code{xg3_variable} column.
It evaluates the combined presence of the involved XG3 variables at each
location.
}
\examples{
\dontrun{
watina <- connect_watina()
library(dplyr)
mylocs <- get_locs(watina, area_codes = "KAL")
mydata <-
 mylocs \%>\%
 get_xg3(watina, 2014)
mydata
eval_xg3_avail(mydata,
               xg3_type = c("L", "V"))
# Disconnect:
DBI::dbDisconnect(watina)
}


}
\seealso{
Other functions to evaluate locations: \code{\link{eval_xg3_series}}
}
\concept{functions to evaluate locations}
