% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chem.R
\encoding{UTF-8}
\name{ggplot_vanwirdum_background}
\alias{ggplot_vanwirdum_background}
\title{Plots hydrochemistry: Van Wirdum diagram}
\usage{
ggplot_vanwirdum_background(
  ir_unit = NULL,
  ec25_unit = "micro",
  contour = "segment",
  lang = "en",
  rhine = FALSE
)
}
\arguments{
\item{ir_unit}{The units for the ionic ratio, can be NULL (default, axis 0-1)
or "pc" (\%, axis 0-100). Choose this parameter according to the unit used
in your dataset.}

\item{ec25_unit}{The units for the electric conductivity et 25°C:
can be "micro" (default, µS/cm) or "milli" (mS/m).
Choose this parameter according to the unit used in your dataset.}

\item{contour}{Draw the mixing contours of the reference water samples,
"segment" (default), "curve" or NULL (do not draw)}

\item{lang}{Which language should be used for the legend, "en" (English,
default) or "nl" (Dutch)}

\item{rhine}{Should the reference point for Rhine be shown? FALSE (default)
or TRUE}
}
\value{
A ggplot object with the Van Wirdum IR-log EC25 diagram, the reference points
for the water types LI-AT-TH and the mixing contours between the reference
water samples LI-AT-TH.
}
\description{
Creates the background of a Van Wirdum diagram for water samples
in ggplot (ionic ratio - log electric conductivity).
}
\details{
Source: Van Wirdum, Geert (1991). Vegetation and hydrology of floating
rich-fens. Datawyse, Maastricht. 316 p. ISBN 90-5291-045-6. (Appendix D)

Creates a ggplot object of the ionic ratio (IR) as Y axis against the
electric conductivity at 25°C (EC25) as X axis and adds reference data.
}
\section{Reference data}{


Reference points (Van Wirdum 1991): benchmark water samples for:

\enumerate{
 \item{lithotrophic water LI: a calcium-bicarbonate type of water, usually
owing its characteristic composition to a contact with soil;}
 \item{atmotrophic water AT: a type of water with low concentrations of most
constituents, usually owing its characteristic composition to atmospheric
precipitation;}
 \item{thalassotrophic water TH: a saline sodium-chloride type of water as
found in the oceans;}
 \item{molunotrophic water RH: polluted water as presently found in the Rhine.}
 }

You can also show the mixing contours between the reference points LI, AT
and TH as curves or as lines. The curved contour encloses the
plotting area of all possible, simple mixtures of the reference water
samples LI-ANG (a relatively calcium-rich groundwater sample ), AT-WTV
(a precipitation sample caught in a relatively unpolluted inland area of
The Netherlands) and TH-N70 (a representative analysis from the North Sea
monitoring program, 70 km from the coast).
Most water analyses plot within the area bounded by the (curved) lines
LI-AT-TH-LI.
}

\section{Typical way of using}{


Add your own water samples as data points (and any other information you
would like to plot) to the Van Wirdum diagram as you would do for any ggplot.
}

\section{Input format}{


Input: a dataset with the electric conductivity at 25°C and the ionic ratio:

 \itemize{
 \item{EC at 25°C can be in µS/cm or mS/m and will be shown on a logarithmic scale}
 \item{IR can be without units (0-1) or in \%}
 }

Compute the ionic ratio as follows:

ionic ratio = [Ca2+] / ([Ca2+] + [Cl-]) with the Ca and Cl concentrations in meq/l

\preformatted{
mydata <- mydata \%>\%
  mutate(Ca_meq = (Ca_mg*2)/40.078,
         Cl_meq = Cl_mg/35.453) \%>\%
  mutate(ir = Ca_meq/(Ca_meq + Cl_meq)) # ir without units (0-1)
}

  or use the helper function \code{\link{calculate_ir}} to calculate the ionic
  ratio based on the Ca and Cl concentrations obtained through
  the \code{\link{get_chem}} function.
}

\examples{
\dontrun{
# a dataset with water samples
mydata <-
  tibble(
    my_ir = runif(10), # without units (0-1)
    my_ec25 = rnorm(10, 100, 35),
    my_site = append(replicate(4, "site A"), replicate(6, "site B"))
  )

# default version of the basis plot without data
ggplot_vanwirdum_background(ec25_unit = "micro",
             contour = "segment",
             lang = "en",
             rhine = FALSE)

# add your own data with EC as x and IR as y and format as you wish
ggplot_vanwirdum_background(contour = "curve",
             lang = "nl",
             rhine = TRUE) +
  geom_point(data = mydata,
             aes(x = my_ec25, y = my_ir, colour = my_site), size = 3) +
  theme(axis.text  = element_text(colour = "blue"))
}

}
