% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{qualify_xg3}
\alias{qualify_xg3}
\title{Qualify XG3 data per location x hydroyear}
\usage{
qualify_xg3(data, xg3_type)
}
\arguments{
\item{data}{An object returned by \code{\link{get_xg3}}.}

\item{xg3_type}{Character vector of length 1, 2 or 3.
Defines the types of XG3 which are taken to evaluate the hydroyears for
each location.
Specifies the 'X' in 'XG3': either \code{"L"}, \code{"H"} and/or \code{"V"}.}
}
\value{
A \code{tbl_lazy} object or a tibble, with columns \code{xg3_variable}
(character; see Details) and \code{available} (logical) to denote at each
location whether the hydroyear has the requested XG3 values available.
}
\description{
Helper function to select the specified XG3 columns and determine
their availability per location x hydroyear.
}
\details{
The column \code{xg3_variable} in the resulting tibble
lists the requested XG3 types, for each location and hydroyear.
If more than one XG3 type is requested, or if \code{data} contains both
vertical CRSes, an extra value "\code{combined}" is listed in the
\code{xg3_variable} column.
It evaluates the combined presence of the involved XG3 variables at each
location and hydroyear.
}
\keyword{internal}
