% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watina.R
\docType{package}
\name{watina-package}
\alias{watina}
\alias{watina-package}
\title{watina: Querying & Processing Data From The INBO Watina Database (mainly groundwater data)}
\description{
The R-package watina contains functions to query
    and process data from the Watina database at the Research Institute for
    Nature and Forest (INBO). This database provides
    groundwater level and chemical data, mainly from natural 
    areas in Flanders (Belgium).
}
\details{
See the README on
\href{https://github.com/inbo/watina/blob/master/README.md}{GitHub}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/inbo/watina}
}

}
\author{
\strong{Maintainer}: Floris Vanderhaeghe \email{floris.vanderhaeghe@inbo.be} (0000-0002-6378-6229)

}
\keyword{internal}
