% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watina.R
\docType{package}
\name{watina-package}
\alias{watina}
\alias{watina-package}
\title{watina: Querying & Processing Data From The INBO Watina Database (mainly groundwater data)}
\description{
The R-package watina contains functions to query
    and process data from the Watina database at the Research Institute for
    Nature and Forest (INBO). This database provides
    groundwater level and chemical data, mainly from natural 
    areas in Flanders (Belgium).
}
\details{
See the \href{00Index.html}{Index} for an overview of the available functions.
Basic data retrieval functions start with "\code{get_}".

Read the vignettes (linked by the \href{00Index.html}{Index}) to get
started with the package.

\code{\link{selectlocs_xg3}} is a top-level function that allows to select
locations based on user-specified conditions on XG3 availability and XG3 series.

\code{\link{selectlocs_chem}} is a top-level function that allows to select
locations based on user-specified conditions on chemical data availability and properties.

For contributing, see the README on
\href{https://github.com/inbo/watina/blob/master/README.md}{GitHub}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://inbo.github.io/watina}
  \item \url{https://github.com/inbo/watina}
  \item Report bugs at \url{https://github.com/inbo/watina/issues}
}

}
\author{
\strong{Maintainer}: Floris Vanderhaeghe \email{floris.vanderhaeghe@inbo.be} (\href{https://orcid.org/0000-0002-6378-6229}{ORCID})

Other contributors:
\itemize{
  \item Peter Desmet \email{peter.desmet@inbo.be} (\href{https://orcid.org/0000-0002-8442-8025}{ORCID}) [contributor]
  \item Jan Wouters \email{jan.wouters@inbo.be} (\href{https://orcid.org/0000-0001-8200-9102}{ORCID}) [contributor]
  \item Frederic Piesschaert \email{frederic.piesschaert@inbo.be} (\href{https://orcid.org/0000-0002-5843-646X}{ORCID}) [contributor]
  \item Mathias Wackenier \email{mathias.wackenier@inbo.be} (\href{https://orcid.org/0000-0001-6934-9241}{ORCID}) [contributor]
  \item Research Institute for Nature and Forest (INBO) \email{info@inbo.be} [copyright holder]
}

}
\keyword{internal}
